refactor(#87): 장비 공용 타입을 equipment 엔티티로 분리
