+++ src/entities/inventory/lib/equipable.ts
@@ -1,5 +1,5 @@
-import type { EquipmentSlot } from "@/entities/dashboard/model/types";
-import { EQUIPMENT_SLOTS } from "@/entities/dashboard/model/types";
+import type { EquipmentSlot } from "@/entities/equipment/model/types";
+import { EQUIPMENT_SLOTS } from "@/entities/equipment/model/types";
 import type { InventoryItemSlot } from "@/entities/inventory/model/types";
 
 const EQUIPPABLE_SLOTS = new Set<EquipmentSlot>(EQUIPMENT_SLOTS);
